[CV-423] Fixed native app running on web, moved processCondoUnits and propertyAdapter to native folder
