+++ apps/native/components/PropertyAdapter.tsx
