+++ package.json
@@ -13,6 +13,9 @@
     "ios": "cd apps/native && npm run ios"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.24.5",
+    "@babel/preset-react": "^7.24.1",
+    "@babel/preset-typescript": "^7.24.1",
     "@testing-library/react-native": "^12.4.3",
     "@types/jest": "^29.5.12",
     "@types/react-datepicker": "^6.2.0",
@@ -22,7 +25,9 @@
     "prettier": "^3.1.1",
     "react-test-renderer": "^18.2.0",
     "ts-jest": "^29.1.2",
-    "turbo": "^1.11.3"
+    "ts-loader": "^9.5.1",
+    "turbo": "^1.11.3",
+    "typescript": "^5.4.5"
   },
   "packageManager": "npm@10.4.0",
   "engines": {
